refactor: extract remaining view-level duplication (scroll anchor, motion, hero)
Continues thinning the views:

- About.tsx and Projects.tsx each open-coded the identical
  searchParams-driven scroll-to-section-or-top effect. Extracted
  useScrollToSearchAnchor so both pages depend on one hook instead of
  duplicating the DOM side effect inline.
- ProjectDetail.tsx defined its own local EASE_OUT + motion props for
  its card entrance, separate from the scrollReveal/mountReveal
  helpers already shared by About/Projects. Added riseIn/pageEnter to
  motion-reveal.ts and switched ProjectDetail.tsx and Home.tsx to use
  it, removing three more inline animation-curve definitions.
- Home.tsx rendered its entire hero (photo, typewriter, heading,
  description, CTA row) inline - the one view untouched by the
  previous SRP pass. Extracted HomeHero (narrow name/description
  props, not the whole user object) so Home.tsx is now just
  Header + HomeHero + Stats.

Verified via tsc, next build, and Playwright checks of all four pages.
